catalog: move backend system out of alpha
