+++ .changeset/four-buttons-draw.md
@@ -1,0 +1,6 @@
+---
+'@backstage/plugin-catalog-backend': patch
+'@backstage/plugin-catalog-node': patch
+---
+
+Deprecated backend system features on the /alpha export, and introduced them on the regular non-alpha export instead, now that the backend system is out of alpha
