+++ docs/backend-system/building-backends/08-migrating.md
@@ -577,7 +577,7 @@ depends on the appropriate extension point and interacts with it.
 
 ```ts title="packages/backend/src/index.ts"
 /* highlight-add-start */
-import { catalogProcessingExtensionPoint } from '@backstage/plugin-catalog-node/alpha';
+import { catalogProcessingExtensionPoint } from '@backstage/plugin-catalog-node';
 import { createBackendModule } from '@backstage/backend-plugin-api';
 /* highlight-add-end */
