+++ plugins/catalog-backend/src/deprecated.ts
@@ -165,3 +165,40 @@ export type AnalyzeOptions = _AnalyzeOptions;
  * @deprecated import from `@backstage/plugin-catalog-node` instead
  */
 export type ScmLocationAnalyzer = _ScmLocationAnalyzer;
+
+import {
+  DefaultCatalogCollatorFactory as _DefaultCatalogCollatorFactory,
+  defaultCatalogCollatorEntityTransformer as _defaultCatalogCollatorEntityTransformer,
+} from '@backstage/plugin-search-backend-module-catalog';
+
+/**
+ * @public
+ * @deprecated import from `@backstage/plugin-search-backend-module-catalog` instead
+ */
+export const DefaultCatalogCollatorFactory = _DefaultCatalogCollatorFactory;
+
+/**
+ * @public
+ * @deprecated import from `@backstage/plugin-search-backend-module-catalog` instead
+ */
+export const defaultCatalogCollatorEntityTransformer =
+  _defaultCatalogCollatorEntityTransformer;
+
+import type {
+  DefaultCatalogCollatorFactoryOptions as _DefaultCatalogCollatorFactoryOptions,
+  CatalogCollatorEntityTransformer as _CatalogCollatorEntityTransformer,
+} from '@backstage/plugin-search-backend-module-catalog';
+
+/**
+ * @public
+ * @deprecated import from `@backstage/plugin-search-backend-module-catalog` instead
+ */
+export type DefaultCatalogCollatorFactoryOptions =
+  _DefaultCatalogCollatorFactoryOptions;
+
+/**
+ * @public
+ * @deprecated import from `@backstage/plugin-search-backend-module-catalog` instead
+ */
+export type CatalogCollatorEntityTransformer =
+  _CatalogCollatorEntityTransformer;
