+++ plugins/catalog-backend/src/service/index.ts
@@ -19,3 +19,4 @@ export type {
   CatalogPermissionRuleInput,
 } from './CatalogBuilder';
 export { CatalogBuilder } from './CatalogBuilder';
+export { catalogPlugin as default } from './CatalogPlugin';
